refactor(make:factory): split command with DefaultPropertiesGuesser
