+++ src/Bundle/Maker/Factory/DefaultPropertiesGuesser.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Zenstruck\Foundry\Bundle\Maker\Factory;
+
+/**
+ * @internal
+ */
+interface DefaultPropertiesGuesser
+{
+    public function __invoke(MakeFactoryData $makeFactoryData, bool $allFields): void;
+
+    public function supports(bool $persisted): bool;
+}
