+++ src/Bundle/Maker/Factory/MakeFactoryPHPDocMethod.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Zenstruck\Foundry\Bundle\Maker;
+namespace Zenstruck\Foundry\Bundle\Maker\Factory;
 
 /**
  * @internal
