+++ src/Test/GlobalStateRegistry.php
@@ -1,7 +1,5 @@
 <?php
 
-declare(strict_types=1);
-
 namespace Zenstruck\Foundry\Test;
 
 use Zenstruck\Foundry\Story;
